+++ packages/stylelint-config-wordpress/__tests__/selectors-scss.test.js
@@ -45,9 +45,9 @@ describe( 'flags warnings with invalid selectors scss', () => {
 		) );
 	});
 
-	it( 'flags four warnings', () => {
+	it( 'flags six warnings', () => {
 		return result.then( data => (
-			expect( data.results[0].warnings ).toHaveLength( 4 )
+			expect( data.results[0].warnings ).toHaveLength( 6 )
 		) );
 	});